+++ abacusflow-usecase/abacusflow-usecase-product/src/main/kotlin/org/bruwave/abacusflow/usecase/product/BasicProductInstanceTO.kt
@@ -1,12 +1,13 @@
 package org.bruwave.abacusflow.usecase.product
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class BasicProductInstanceTO(
     val id: Long,
     val name: String,
     val serialNumber: String,
-    val unitPrice: Double,
+    val unitPrice: BigDecimal,
     val productId: Long,
     val productName: String,
     val purchaseOrderNo: UUID,
@@ -17,6 +18,6 @@ data class ProductInstanceForBasicProductTO(
     val id: Long,
     val name: String,
     val serialNumber: String,
-    val unitPrice: Double,
+    val unitPrice: BigDecimal,
     val productId: Long,
 )
